Applied refactoring
Moved helper function to a dedicated module.
Fixed tests.
