+++ tests/conftest.py
@@ -2,24 +2,23 @@ import pytest
 
 
 @pytest.fixture
-def small_bricks_wall():
+def small_fixture():
+    """Provide a brick wall structure and bricks count"""
+
     return [[1], [1], [1]], 3
 
 
 @pytest.fixture
-def bricks_wall():
+def regular_fixture():
+    """Provide a brick wall structure and bricks count"""
+
     return [
         [5, 5, 3, 4, 1],
         [3, 2, 1, 4, 2, 1, 5],
         [2, 3, 1, 5, 5, 2],
         [3, 4, 3, 4, 3, 1],
         [5, 5, 3, 2, 3],
-    ]
-
-
-@pytest.fixture
-def number_of_bricks():
-    return 29
+    ], 29
 
 
 @pytest.fixture
